+++ src/main/java/searchengine/repositories/PageRepository.java
@@ -8,7 +8,7 @@ import searchengine.model.Site;
 import java.util.Optional;
 
 @Repository
-public interface PageRepository extends JpaRepository<Page, Long> {
+public interface PageRepository extends JpaRepository<Page, Integer> {
     Optional<Page> findByPathAndSite(String path, Site site);
     boolean existsByPathAndSite(String path, Site site);
     Integer countAllBySite(Site site);
